requirements.txt
pandas==0.23.4
numpy==1.16.2
PyYAML==5.1.2
